+++ ricaun.NUnit/Extensions/ReferenceLoaderUtils.NETFRAMEWORK.cs
@@ -11,27 +11,32 @@
     /// </summary>
     internal static partial class ReferenceLoaderUtils
     {
+        internal class AppDomainDisposable : IDisposable
+        {
+            private readonly AppDomain _appDomain;
+            public AppDomain AppDomain => _appDomain;
+            public AppDomainDisposable(string assemblyPath = null)
+            {
+                var settings = new AppDomainSetup
+                {
+                    ApplicationBase = AppDomain.CurrentDomain.BaseDirectory
+                };
+                _appDomain = AppDomain.CreateDomain(Guid.NewGuid().ToString(), null, settings);
+            }
+            public void Dispose()
+            {
+                AppDomain.Unload(_appDomain);
+            }
+        }
+
         public static AssemblyMetadataAttribute[] GetAssemblyMetadataAttributes(string assemblyPath)
         {
-            //var result = new AssemblyMetadataAttribute[] { };
-
-            //var assembly = Assembly.Load(File.ReadAllBytes(assemblyPath));
-            //result = assembly.GetCustomAttributes<AssemblyMetadataAttribute>().ToArray();
-
-            var settings = new AppDomainSetup
+            using (var domain = new AppDomainDisposable())
             {
-                ApplicationBase = AppDomain.CurrentDomain.BaseDirectory,
-            };
-            var childDomain = AppDomain.CreateDomain(Guid.NewGuid().ToString(), null, settings);
-
-            var loader = childDomain.CreateReferenceLoader();
-
-            //This operation is executed in the new AppDomain
-            var result = loader.GetAssemblyMetadataAttributes(assemblyPath);
-
-            AppDomain.Unload(childDomain);
-
-            return result.Select(e=>e.GetAssemblyMetadataAttribute()).ToArray();
+                var loader = domain.AppDomain.CreateReferenceLoader();
+                var result = loader.GetAssemblyMetadataAttributes(assemblyPath);
+                return result.Select(e => e.GetAssemblyMetadataAttribute()).ToArray();
+            }
         }
 
         /// <summary>
@@ -41,20 +46,11 @@
         /// <returns></returns>
         public static AssemblyName[] GetReferencedAssemblies(string assemblyPath)
         {
-            var settings = new AppDomainSetup
+            using (var domain = new AppDomainDisposable())
             {
-                ApplicationBase = AppDomain.CurrentDomain.BaseDirectory,
-            };
-            var childDomain = AppDomain.CreateDomain(Guid.NewGuid().ToString(), null, settings);
-
-            var loader = childDomain.CreateReferenceLoader();
-
-            //This operation is executed in the new AppDomain
-            var assemblyNames = loader.LoadReferences(assemblyPath);
-
-            AppDomain.Unload(childDomain);
-
-            return assemblyNames;
+                var loader = domain.AppDomain.CreateReferenceLoader();
+                return loader.LoadReferences(assemblyPath);
+            }
         }
 
         private static ReferenceLoader CreateReferenceLoader(this AppDomain domain)
@@ -110,29 +106,19 @@
                 return assemblyNames;
             }
 
+            /// <summary>
+            /// Get the metadata attributes of the specified assembly.
+            /// </summary>
+            /// <param name="assemblyPath">The path of the assembly.</param>
+            /// <returns>An array of AssemblyMetadataSerializable objects representing the metadata attributes.</returns>
+            /// <remarks><see cref="AssemblyMetadataAttribute"/> can not be use because does not have <see cref="SerializableAttribute"/></remarks>
             public AssemblyMetadataSerializable[] GetAssemblyMetadataAttributes(string assemblyPath)
             {
                 var assembly = Assembly.LoadFrom(assemblyPath);
                 return assembly.GetCustomAttributes<AssemblyMetadataAttribute>()
-                    .Select(e => new AssemblyMetadataSerializable(e))
+                    .Select(AssemblyMetadataSerializable.Create)
                     .ToArray();
             }
-        }
-
-        [Serializable]
-        public class AssemblyMetadataSerializable
-        {
-            public AssemblyMetadataSerializable(AssemblyMetadataAttribute assemblyMetadataAttribute)
-            {
-                Key = assemblyMetadataAttribute.Key;
-                Value = assemblyMetadataAttribute.Value;
-            }
-            public AssemblyMetadataAttribute GetAssemblyMetadataAttribute()
-            {
-                return new AssemblyMetadataAttribute(Key, Value);
-            }
-            public string Key { get; }
-            public string Value { get; }
         }
     }
 #endif